Change port for production server config
I get the error "listen EACCES" when running "npm run start" on port 80.

According to this [SO post](http://stackoverflow.com/questions/18947356/node-js-app-cant-run-on-port-80-even-though-theres-no-other-process-blocking-t), I needed to run the production server under a different port.

Not sure if you want to change the port number here to make "npm run start" work or add a note to the README production step to tell people that the error they see is related to a permission problem and not something in the server.js file. Let me know if I understand this error correctly. Thanks.